+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Slightly improve map load time via more aggressive chunk preloading.
 - Improve style parsing performance.
+- Preload Meshopt decoder earlier to improve 3D styles load time.
 
 ### Bug fixes 🐞